show only projects of the current user
refs #36

+++ src/delegates/projects.delegate.js
@@ -3,6 +3,8 @@
 // functionality related to HITs.
 
 const projectsDao = require(__base + 'dao/projects.dao');
+const usersDao = require(__base + 'dao/users.dao');
+
 //error handler
 const errHandler = require(__base + 'utils/errors');
 //supply the auxiliary function
@@ -15,10 +17,15 @@ const validationSchemes = require(__base + 'utils/validation.schemes');
 
 /**
  * insert a project
+ * @param {string} tokenId of user
  * @param {object} newProjectData
  * @returns {object} project created
  */
-async function insert(newProjectData) {
+async function insert(tokenId, newProjectData ) {
+
+    //error check for tokenId
+    support.isValidTokenId(tokenId);
+
     //check input format
     let valid = ajv.validate(validationSchemes.project, newProjectData);
     //if is not a valid input
@@ -26,6 +33,15 @@ async function insert(newProjectData) {
     {
         throw errHandler.createBadRequestError('the new project data is not valid!');
     }
+    //get user info
+    let user = await usersDao.getUserByTokenId(tokenId);
+    //if user does not exist
+    if(!user){
+        throw errHandler.createBadRequestError("the token does not match any user!");
+    }
+    //add the user_id in project data
+    newProjectData.user_id = user.id;
+
 
     //call DAO layer
     let res = await projectsDao.insert(newProjectData);
@@ -37,12 +53,15 @@ async function insert(newProjectData) {
 /**
  *  * update a project
  * @param {int}  project_id
+ * @param {string} tokenId of user
  * @param {object} newProjectData
  */
-async function update(project_id, newProjectData) {
+async function update(project_id, tokenId, newProjectData) {
 
     //check validation of project id and transform the value in integer
     project_id = support.setAndCheckValidProjectId(project_id);
+    //error check for tokenId
+    support.isValidTokenId(tokenId);
 
     //check input format
     let valid = ajv.validate(validationSchemes.project, newProjectData);
@@ -50,6 +69,22 @@ async function update(project_id, newProjectData) {
     if (!valid)
     {
         throw errHandler.createBadRequestError('the new project data for update is not valid!');
+    }
+
+    //get user info
+    let user = await usersDao.getUserByTokenId(tokenId);
+    //if user does not exist
+    if(!user){
+        throw errHandler.createBadRequestError("the token does not match any user!");
+    }
+    //add the user_id in project data
+    newProjectData.user_id = user.id;
+
+    //check relationship between the project and user
+    let project = await projectsDao.selectByIdAndUserId(project_id, user.id);
+    //if the user isn't project's owner
+    if(!project){
+        throw errHandler.createUnauthorizedError("unauthorized operation");
     }
 
     //call DAO layer
@@ -66,11 +101,28 @@ async function update(project_id, newProjectData) {
 /**
  *  * delete a project
  * @param {int} project_id
+ * @param {string} tokenId of user
  */
-async function deletes(project_id) {
+async function deletes(project_id, tokenId) {
 
     //check validation of project id and transform the value in integer
     project_id = support.setAndCheckValidProjectId(project_id);
+    //error check for tokenId
+    support.isValidTokenId(tokenId);
+
+    //get user info
+    let user = await usersDao.getUserByTokenId(tokenId);
+    //if user does not exist
+    if(!user){
+        throw errHandler.createBadRequestError("the token does not match any user!");
+    }
+    //check relationship between the project and user
+    let project = await projectsDao.selectByIdAndUserId(project_id, user.id);
+    //if the user isn't project's owner
+    if(!project){
+        throw errHandler.createUnauthorizedError("unauthorized operation");
+    }
+
 
     //call DAO layer
     let numberRow = await projectsDao.deletes(project_id);
@@ -85,17 +137,33 @@ async function deletes(project_id) {
 /**
  * select a project
  * @param {int} project_id
+ * @param {string} tokenId of user
  * @returns {object} project found
  */
-async function selectById(project_id) {
+async function selectById(project_id, tokenId) {
 
     //check validation of project id and transform the value in integer
     project_id = support.setAndCheckValidProjectId(project_id);
+    //error check for tokenId
+    support.isValidTokenId(tokenId);
+
+    //get user info
+    let user = await usersDao.getUserByTokenId(tokenId);
+    //if user does not exist
+    if(!user){
+        throw errHandler.createBadRequestError("the token does not match any user!");
+    }
+    //check relationship between the project and user
+    let project = await projectsDao.selectByIdAndUserId(project_id, user.id);
+    //if the user isn't project's owner
+    if(!project){
+        throw errHandler.createUnauthorizedError("unauthorized operation");
+    }
 
     //call DAO layer
     let res = await projectsDao.selectById(project_id);
     //error check
-    if (res === undefined)
+    if (!res)
     {
         throw errHandler.createNotFoundError('Project does not exist!');
     }
@@ -105,8 +173,8 @@ async function selectById(project_id) {
 /**
  * 
  * select all project
- * @param {string} orderBy [id, date_created, date_last_modified, date_deleted}
- * @param {string} sort {ASC or DESC}
+ * @param {string} orderBy [id, date_created, date_last_modified, date_deleted]
+ * @param {string} sort [ASC or DESC]
  * @param {int} start offset position where we begin to get
  * @param {int} count number of projects
  * @returns {Object} array of projects and total number of result
@@ -129,6 +197,43 @@ async function selectAll(orderBy, sort, start, count) {
     return res;
 }
 
+/**
+ * select all project of a specific user
+ * @param {string} tokenId of user
+ * @param {string} orderBy [id, date_created, date_last_modified, date_deleted]
+ * @param {string} sort [ASC or DESC]
+ * @param {int} start offset position where we begin to get
+ * @param {int} count number of projects
+ * @returns {Object} array of projects and total number of result
+ */
+async function selectAllByUserId(tokenId, orderBy, sort, start, count) {
+
+    //error check for tokenId
+    support.isValidTokenId(tokenId);
+
+    //check the validation of parameters
+    orderBy = support.setAndCheckValidProjectOrderBy(orderBy);
+    sort = support.setAndCheckValidSort(sort);
+    start = support.setAndCheckValidStart(start);
+    count = support.setAndCheckValidCount(count);
+
+    //get user info
+    let user = await usersDao.getUserByTokenId(tokenId);
+    //if user does not exist
+    if(!user){
+        throw errHandler.createBadRequestError("the token does not match any user!");
+    }
+
+    let res = await projectsDao.selectAllByUserId(user.id, orderBy, sort, start, count);
+
+    //error check
+    if (res.results.length === 0)
+    {
+        throw errHandler.createNotFoundError('the list is empty!');
+    }
+    return res;
+}
+
 
 /**
  * select project by a single keyword
@@ -138,7 +243,7 @@ async function selectAll(orderBy, sort, start, count) {
  * @param {int} start offset position where we begin to get
  * @param {int} count number of projects
  * @returns {Object} array of projects and total number of result
- */
+ *//*
 async function selectBySingleKeyword(keyword, orderBy, sort, start, count) {
 
 
@@ -157,7 +262,7 @@ async function selectBySingleKeyword(keyword, orderBy, sort, start, count) {
         throw errHandler.createNotFoundError('the list is empty!');
     }
     return res;
-}
+}*/
 
 
 
@@ -170,6 +275,7 @@ module.exports = {
     deletes,
     selectById,
     selectAll,
-    selectBySingleKeyword,
+    selectAllByUserId,
+    //selectBySingleKeyword,
 
 };
